# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "A RESTful API for managing news articles and user preferences, built with Laravel 11, Sanctum authentication, Redis caching, and rate limiting (60 requests/minute per IP).",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Retrieve a list of articles",
                "description": "Returns a paginated list of articles with optional filters for keyword, date, category, and source.",
                "operationId": "d87f751e8f428fb968c2791aa22f017e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "example": 10
                        }
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search keyword to filter articles by title, content, or summary",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "AI"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter by publication date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-02-23"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by article category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "sports"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter by article source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Gazette"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Articles retrieved successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Articles retrieved successfully"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "provider": {
                                                                "type": "string",
                                                                "example": "NewsData.Io"
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "sports"
                                                            },
                                                            "source": {
                                                                "type": "string",
                                                                "example": "Gazette"
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Colorado Springs pet related information and events starting Feb. 22"
                                                            },
                                                            "content": {
                                                                "type": "string",
                                                                "example": "ONLY AVAILABLE IN PAID PLANS"
                                                            },
                                                            "summary": {
                                                                "type": "string",
                                                                "example": "Pet related information and events in and around the Colorado Springs area"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "BY CARLOTTA OLSON The Gazette"
                                                            },
                                                            "article_url": {
                                                                "type": "string",
                                                                "example": "https://gazette.com/events-calendars/..."
                                                            },
                                                            "image_url": {
                                                                "type": "string",
                                                                "example": "https://bloximages.newyork1.vip.townnews.com/..."
                                                            },
                                                            "published_at": {
                                                                "type": "string",
                                                                "example": "2025-02-22"
                                                            },
                                                            "x_id": {
                                                                "type": "string",
                                                                "example": "ME4PbazW6OVB"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles?page=1"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page_url": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles?page=1"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string",
                                                                "example": "&laquo; Previous"
                                                            },
                                                            "active": {
                                                                "type": "boolean",
                                                                "example": false
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://127.0.0.1:8000/api/articles"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized: Invalid or expired token."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Too many requests",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Too many requests. Please try again later."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 429
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}